FIX: Remove addons that fail to download.
Generally speaking, the addons that fail their download are legacy
entries that exist on packages but not github. Deleting them is as such
reasonable.

If the issue is temporary, then these records will be recreated on the
next execution.

Note that if the issue is permanent, these entries will be continually
deleted and recreated. It may be better to set a status flag instead,
but that will need to be passed into elasticsearch and applied to our
default view-list and search queries.