FIX Don't create addons with empty names
Since 2019, this has happened about 100 times without a clear pattern.
This is the only place where addons are created, and the name is never updated,
so this should guard against this case effectively.

I've run a manual DELETE FROM Addons WHERE Name IS NULL on the database already to clear up existing faults.