Predictable default version
Previously it was pretty much random, since it used the default_sort of "ID DESC".
This chooses dev-master most of the time, but is resilient against folks who name their branches differently.
Ideally we could use the default branch configured in Github etc, but that's too much work for now