NEW Packagist API performance
Hopefully avoids us getting rate limited. Uses statically cached files with less metadata,
so we need to shift some of the addon data population from the updater to the builder.
This doesn't reduce the number of API calls overall, but it makes most of them a lot "cheaper" for Packagist.

See https://packagist.org/apidoc:

> Using composer metadata: This is the preferred way to access the data as it is always up to date, and dumped to static files so it is very efficient on our end.

The UI is already set up to deal with partially populated packages