FIX Composer v2 compat, requires install
Previously it was enough to download() a package,
which would actually write files into the provided $targetDir.
The method still claims to do that, but for the GitDownloader
it uses a separate global composer cache instead and doesn't touch $targetDir.
You need to call install() to actually write to the path.

Weirdly, this seems different from a "composer install",
it only copies the files, and doesn't create a vendor/ folder.
Which is desireable in our case, since we don't want the builder
to be slowed down by checking out more files we don't need.