Don't build *all* addons nightly
Instead, rely on UpdateAddonsTask to detect which modules actually require a build.
This increases the chance of missing an update, but it's simply too expensive
to get the metadata from 3000+ modules from a dynamic Packagist API endpoint.
It appears that we're getting rate limited due to over use.

It's also not effective to call the Github API for the markdown builds on each of them.
The (personal) tokens used for this access share the same rate limits as other use cases,
so we need to treat this as a limited resource as well.